Added new permissions ownercomment, groupcomment and publiccomment into wiki
git-svn-id: https://source.sakaiproject.org/svn/rwiki/branches/SAK-8391@65907 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ rwiki-impl/impl/src/java/uk/ac/cam/caret/sakai/rwiki/component/service/impl/BaseFOPSerializer.java
@@ -169,12 +169,12 @@ public class BaseFOPSerializer extends ToXMLSAXHandler implements ContentHandler
 											+ "\nContent-length: "
 											+ resource.getContentLength()
 											+ "\n\n";
-									byte[] output = new byte[headers.length()
-											+ resource.getContentLength()];
+									byte[] output = new byte[(int) (headers.length()
+		+ resource.getContentLength())];
 									System.arraycopy(headers.getBytes(), 0,
 											output, 0, headers.length());
 									System.arraycopy(resource.getContent(), 0,
-											output, headers.length(), resource
+											output, headers.length(), (int) resource
 													.getContentLength());
 									return new StreamSource(resource
 											.streamContent());
